requirements.txt
# Dependências do Detector de Gestos
opencv-python>=4.8.0
mediapipe>=0.10.14
numpy>=1.26.0
zeroconf==0.131.0
flask==3.0.0
requests==2.31.0
pyperclip==1.8.2
Pillow==10.1.0
pywin32==306; sys_platform == 'win32'